Re-organized into single package with sub-directories for interface, implementations and tests
