+++ descartes_light/core/include/descartes_light/core/waypoint_sampler.h
