--- descartes_light/src/ladder_graph/ladder_graph.cpp
@@ -1,11 +1,0 @@
-#include <descartes_light/impl/ladder_graph.hpp>
-#include <vector>
-#include <list>
-
-namespace descartes_light
-{
-// Explicit template specialization
-template class LadderGraph<float>;
-template class LadderGraph<double>;
-
-}  // namespace descartes_light
