+++ descartes_light/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 find_gtest()
 add_executable(${PROJECT_NAME}_utest utest.cpp)
-target_link_libraries(${PROJECT_NAME}_utest PRIVATE ${PROJECT_NAME}_ladder_graph GTest::GTest GTest::Main)
+target_link_libraries(${PROJECT_NAME}_utest PRIVATE ${PROJECT_NAME}_solvers GTest::GTest GTest::Main)
 target_clang_tidy(${PROJECT_NAME}_utest ENABLE ${DESCARTES_ENABLE_CLANG_TIDY} ARGUMENTS ${DESCARTES_CLANG_TIDY_ARGS})
 target_code_coverage(${PROJECT_NAME}_utest ALL ENABLE ${DESCARTES_ENABLE_CODE_COVERAGE})
 add_gtest_discover_tests(${PROJECT_NAME}_utest)
