--- descartes_samplers/cmake/descartes_samplers-config.cmake.in
@@ -1,10 +1,0 @@
-@PACKAGE_INIT@
-
-set(@PROJECT_NAME@_FOUND ON)
-set_and_check(@PROJECT_NAME@_INCLUDE_DIRS "${PACKAGE_PREFIX_DIR}/include")
-set_and_check(@PROJECT_NAME@_LIBRARY_DIRS "${PACKAGE_PREFIX_DIR}/lib")
-
-include(CMakeFindDependencyMacro)
-find_dependency(descartes_light)
-
-include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@-targets.cmake")
