--- descartes_samplers/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>descartes_samplers</name>
-  <version>0.1.0</version>
-  <description>The descartes_samplers package</description>
-  <author>Michael Ripperger</author>
-  <author>Levi Armstrong</author>
-  <author>Jon Meyer</author>
-  <maintainer email="michael.ripperger@swri.org">Michael Ripperger</maintainer>
-  <maintainer email="levi.armstrong@swri.org">Levi Armstrong</maintainer>
-  <license>Apache 2.0</license>
-
-  <!-- Following recommendations of REP 136 -->
-  <exec_depend>catkin</exec_depend>
-  <buildtool_depend>cmake</buildtool_depend>
-
-  <depend>descartes_light</depend>
-  <depend>eigen</depend>
-  <build_depend>ros_industrial_cmake_boilerplate</build_depend>
-
-  <export>
-    <build_type>cmake</build_type>
-  </export>
-</package>
